Added stuff
I just realized that the UP, DWN, PTT buttons should be powered by the radios.

This means that I can get rid of the DAC, and just output to a voltage divider to generate the voltages.

The radios have tons of 47k resistors on these lines so driving a lot of current isn't a requirement